TVC logic setup (#11)
* Create TVC logic.py

Needs more work. Idk how to reference specific elements in the array yet, but it's a start.

* TVC logic quat normalization + Euler

Hopefully added code for normalizing the quaternion, and calculating one of the Euler angles. It should be possible to do a sanity check with this.

* TVC Euler math

Finished Euler angle conversion math

* Create BasicKalmanTest.py

File for Kalman filter testing, no idea if it'll work

* TVC configuration

something something conflicted file idk seems fine to me

* TVC configuration defuck

* stupid fucking radian conversion + format defuck

---------

Co-authored-by: Brandon <68145204+besser435@users.noreply.github.com>